fix signed/unsigned cast

+++ rjit/src/rir/Primitives.cpp
@@ -8,7 +8,7 @@ namespace rir {
 
 namespace {
 
-void compileSpecial(CodeStream& cs, long special_id, num_args_t nargs) {
+void compileSpecial(CodeStream& cs, uintptr_t special_id, num_args_t nargs) {
     if (special_id == Primitives::do_begin_id) {
         // TODO have a loop
         if (nargs == 0) {
@@ -73,7 +73,7 @@ BCClosure* Primitives::compilePrimitive(SEXP fun, num_args_t nargs) {
 
     switch (TYPEOF(fun)) {
     case SPECIALSXP: {
-        long idx = (long)CAR(fun);
+        uintptr_t idx = (uintptr_t)(CAR(fun));
         compileSpecial(cs, idx, nargs);
         break;
     }
